requirements.txt
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
exceptiongroup==1.1.0
h11==0.14.0
idna==3.4
numpy==1.24.1
opencv-python==4.7.0.68
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
pycparser==2.21
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.28.2
selenium==4.8.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
webdriver-manager==3.8.5
wsproto==1.2.0
